# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "paths": {
        "/api/get_companies": {
            "get": {
                "tags": [
                    "公司資訊"
                ],
                "summary": "取得公司資訊",
                "description": "請給我對應的id",
                "operationId": "App\\Http\\Controllers\\api\\CompanyController::get_companies",
                "parameters": [
                    {
                        "name": "works[]",
                        "in": "query",
                        "description": "請輸入查詢id",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "請求成功"
                    }
                }
            }
        },
        "/api/industryCategoryCount": {
            "get": {
                "tags": [
                    "公司資訊"
                ],
                "summary": "取得產業類別資訊",
                "description": "請給我對應的id",
                "operationId": "App\\Http\\Controllers\\api\\CompanyController::getIndustryCategoryCount",
                "parameters": [
                    {
                        "name": "works[]",
                        "in": "query",
                        "description": "請輸入查詢id",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "請求成功"
                    }
                }
            }
        },
        "/api/capital": {
            "get": {
                "tags": [
                    "給我職缺資訊"
                ],
                "summary": "取得資本額",
                "description": "請給我對應的id",
                "operationId": "App\\Http\\Controllers\\api\\CompanyController::getCapital",
                "parameters": [
                    {
                        "name": "works[]",
                        "in": "query",
                        "description": "請輸入查詢id",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "請求成功"
                    }
                }
            }
        },
        "/api/workers": {
            "get": {
                "tags": [
                    "給我職缺資訊"
                ],
                "summary": "取得員工人數",
                "description": "請給我對應的id",
                "operationId": "App\\Http\\Controllers\\api\\CompanyController::getWorkers",
                "parameters": [
                    {
                        "name": "works[]",
                        "in": "query",
                        "description": "請輸入查詢id",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "請求成功"
                    }
                }
            }
        },
        "/api/getVacancies": {
            "get": {
                "tags": [
                    "職缺資訊"
                ],
                "summary": "取得職缺資訊",
                "description": "請給我對應的id",
                "operationId": "App\\Http\\Controllers\\api\\WorkController::getVacancies",
                "parameters": [
                    {
                        "name": "works[]",
                        "in": "query",
                        "description": "請輸入查詢id",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "請求成功"
                    }
                }
            }
        },
        "/api/getCategories": {
            "get": {
                "tags": [
                    "職缺資訊"
                ],
                "summary": "取得職缺種類資訊",
                "description": "請給我對應的id",
                "operationId": "App\\Http\\Controllers\\api\\WorkController::getCategories",
                "parameters": [
                    {
                        "name": "works[]",
                        "in": "query",
                        "description": "請輸入查詢id",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "請求成功"
                    }
                }
            }
        },
        "/api/get_tools": {
            "get": {
                "tags": [
                    "職缺資訊"
                ],
                "summary": "取得職缺需求工具資訊",
                "description": "請給我對應的id",
                "operationId": "App\\Http\\Controllers\\api\\WorkController::getTools",
                "parameters": [
                    {
                        "name": "works[]",
                        "in": "query",
                        "description": "請輸入查詢id",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "請求成功"
                    }
                }
            }
        },
        "/api/categoryCount": {
            "get": {
                "tags": [
                    "職缺資訊"
                ],
                "summary": "取得種類統計分析資訊",
                "description": "請給我對應的id",
                "operationId": "App\\Http\\Controllers\\api\\WorkController::getCategoryCount",
                "parameters": [
                    {
                        "name": "works[]",
                        "in": "query",
                        "description": "請輸入查詢id",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "請求成功"
                    }
                }
            }
        },
        "/api/claimEducationCount": {
            "get": {
                "tags": [
                    "職缺資訊"
                ],
                "summary": "取得需求學歷分析資訊",
                "description": "請給我對應的id",
                "operationId": "App\\Http\\Controllers\\api\\WorkController::getVacancyClaimEducationCount",
                "parameters": [
                    {
                        "name": "works[]",
                        "in": "query",
                        "description": "請輸入查詢id",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "請求成功"
                    }
                }
            }
        },
        "/api/claimExperienceCount": {
            "get": {
                "tags": [
                    "職缺資訊"
                ],
                "summary": "取得工作經歷分析資訊",
                "description": "請給我對應的id",
                "operationId": "App\\Http\\Controllers\\api\\WorkController::getVacancyClaimExperienceCount",
                "parameters": [
                    {
                        "name": "works[]",
                        "in": "query",
                        "description": "請輸入查詢id",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "請求成功"
                    }
                }
            }
        },
        "/api/toolCount": {
            "get": {
                "tags": [
                    "職缺資訊"
                ],
                "summary": "取得工具統計分析資訊",
                "description": "請給我對應的id",
                "operationId": "App\\Http\\Controllers\\api\\WorkController::getToolCount",
                "parameters": [
                    {
                        "name": "works[]",
                        "in": "query",
                        "description": "請輸入查詢id",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "請求成功"
                    }
                }
            }
        }
    }
}